+++ timberborn_platform_cruncher_gui/src/app.rs
@@ -1,18 +1,15 @@
-use std::{
-    cmp::Ordering,
-    collections::{BTreeMap, BTreeSet},
-    ops::ControlFlow,
-};
+use std::ops::ControlFlow;
 
 #[allow(unused_imports)] // Keeping this Anyhow import as Context would clash with egui
 use anyhow::Context as _;
 use eframe::Frame;
 use egui::{
     Button, Color32, Context, DragValue, Modal, PointerButton, Rect, Response, RichText, Sense,
-    Stroke, StrokeKind, TextEdit, Ui, UiBuilder, Vec2, Widget, pos2, vec2,
+    Stroke, StrokeKind, Ui, UiBuilder, Vec2, Widget, pos2, vec2,
 };
 use itertools::Itertools;
 use log::{error, info};
+use platform_type_selector::PlatformTypeSelector;
 use rustsat::solvers::{Interrupt, Solve, SolveStats, SolverResult};
 use timberborn_platform_cruncher::{
     encoder::{Encoding, PlatformLayout, PlatformLimits},
@@ -25,6 +22,7 @@ use timberborn_platform_cruncher::{
 use crate::{SolverBackend, app::frame_history::FrameHistory};
 
 mod frame_history;
+mod platform_type_selector;
 
 #[derive(Clone, Default, Debug)]
 struct TerrainTile {
@@ -247,156 +245,6 @@ where
                 self.start_solver(Default::default());
             }
         });
-    }
-}
-
-struct PlatformTypeSelector {
-    platform_defs: BTreeMap<PlatformDefOrdered, PlatformDefItemData>,
-    new_platform_str: String,
-    focus_text_next_frame: bool,
-    text_box_feedback: Option<TextBoxFeedbackKind>,
-}
-
-#[derive(Clone, Debug)]
-enum TextBoxFeedbackKind {
-    Error,
-    Warning,
-}
-
-impl TextBoxFeedbackKind {
-    pub fn as_color(&self) -> Color32 {
-        match self {
-            TextBoxFeedbackKind::Error => Color32::RED,
-            TextBoxFeedbackKind::Warning => Color32::ORANGE,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, Hash)]
-struct PlatformDefOrdered(pub PlatformDef);
-
-#[derive(Clone, Debug)]
-struct PlatformDefItemData {
-    active: bool,
-}
-
-impl PartialEq for PlatformDefOrdered {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.dims().eq(&other.0.dims())
-    }
-}
-
-impl Eq for PlatformDefOrdered {}
-
-impl Ord for PlatformDefOrdered {
-    fn cmp(&self, other: &Self) -> Ordering {
-        let self_dims = self.0.dims();
-        let other_dims = other.0.dims();
-        match (self_dims.width.cmp(&other_dims.width), self_dims.height.cmp(&other_dims.height)) {
-            (Ordering::Equal, height_cmp) => height_cmp,
-            (width_cmp, _) => width_cmp,
-        }
-    }
-}
-
-impl PartialOrd for PlatformDefOrdered {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PlatformDefItemData {
-    pub fn new_active() -> Self {
-        Self { active: true }
-    }
-}
-
-impl PlatformTypeSelector {
-    pub fn with_defaults(
-        platform_defs: impl IntoIterator<Item = PlatformDef>,
-    ) -> PlatformTypeSelector {
-        let platform_defs_btree = platform_defs
-            .into_iter()
-            .map(|def| (PlatformDefOrdered(def), PlatformDefItemData::new_active()))
-            .collect();
-        dbg!(&platform_defs_btree);
-        Self {
-            platform_defs: platform_defs_btree,
-            new_platform_str: String::new(),
-            focus_text_next_frame: false,
-            text_box_feedback: None,
-        }
-    }
-
-    pub fn ui(&mut self, ui: &mut Ui) {
-        ui.vertical(|ui| {
-            // Rows with all platforms
-            // Items are removed by returning false from the closure
-            &mut self.platform_defs.retain(|def, data| {
-                let dims = def.0.dims();
-                // Disable removal/deactivation for 1x1 platforms,
-                // as a lot of logic assumes that they're always available
-                let disable_controls = dims != Dimensions::new(1, 1);
-                ui.add_enabled_ui(disable_controls, |ui| {
-                    ui.horizontal(|ui| {
-                        let should_keep = !ui.button("X").clicked();
-                        ui.checkbox(&mut data.active, egui::Atom::default());
-                        ui.label(format!("{}x{}", dims.width(), dims.height()));
-
-                        should_keep
-                    })
-                    .inner
-                })
-                .inner
-            });
-
-            // Last row with an entry box
-            ui.horizontal(|ui| {
-                let plus_clicked = ui.button("+").clicked();
-
-                let text_edit = TextEdit::singleline(&mut self.new_platform_str)
-                    .text_color_opt(self.text_box_feedback.as_ref().map(|x| x.as_color()))
-                    .show(ui);
-
-                if self.focus_text_next_frame {
-                    self.focus_text_next_frame = false;
-                    text_edit.response.request_focus();
-                }
-
-                if text_edit.response.lost_focus() || text_edit.response.changed() {
-                    self.text_box_feedback = None;
-                }
-
-                if plus_clicked
-                    || text_edit.response.lost_focus()
-                        && ui.input(|i| i.key_pressed(egui::Key::Enter))
-                {
-                    if let Some(def) = try_parse_platform_def(&self.new_platform_str) {
-                        if self
-                            .platform_defs
-                            .insert(PlatformDefOrdered(def), PlatformDefItemData::new_active())
-                            .is_none()
-                        {
-                            dbg!(&self.platform_defs);
-                            self.new_platform_str.clear();
-                        } else {
-                            self.text_box_feedback = Some(TextBoxFeedbackKind::Warning);
-                        }
-                    } else {
-                        info!("Failed to parse platform definition");
-                        self.text_box_feedback = Some(TextBoxFeedbackKind::Error);
-                    }
-
-                    self.focus_text_next_frame = true;
-                }
-            });
-        });
-    }
-
-    pub fn active_platform_defs(&self) -> impl Iterator<Item = PlatformDef> {
-        self.platform_defs.iter().filter_map(|(def, &PlatformDefItemData { active, .. })| {
-            active.then_some(def.0.clone())
-        })
     }
 }
